requirements.txt
fastapi==0.114.1
uvicorn==0.30.6
pydantic==2.9.1
python-telegram-bot[job-queue]==22.5
